Double clicks working
Added in code to check all the touch events at the same time instead of
one at a time. This allows for a touch on the right box and a tap on the
left side of the box to both increase the counter and holding the right
and tapping the left increases the left multiple times.